server/requirements.txt
djangorestframework>=3.13.1
Django>=3.2.6
gunicorn>=20.1.0
psycopg2-binary>=2.9.1
pyjwt>=2.4.0
django-cors-headers>=3.12.0
drf-yasg>=1.20.0